button functionality working
