+++ app/src/pages/Compress.tsx
@@ -4,16 +4,18 @@ import ToolInterface from "../components/ToolInterface";
 
 export default function Compress() {
   const [isReady, setIsReady] = useState<boolean>(false);
-  const [file, setFile] = useState<File>() as [File, () => void]
-  const handleClick = async() => {
-    //make different logic for when isReady is true(download logic)
+  const [file, setFile] = useState<File>() as [File, () => void];
 
-  }
   return (
     <div>
       <div className="h-full flex flex-col justify-center items-center">
-        <ToolInterface fileState={{file, setFile}} heading="Compress" />
-        <Button onClick={handleClick}  isReady={isReady}/>
+        <ToolInterface fileState={{ file, setFile }} heading="Compress" />
+        <Button
+          file={file}
+          isReady={isReady}
+          tool="compress"
+          setIsReady={setIsReady}
+        />
       </div>
     </div>
   );
